pull snapshots out into their own helper
